+++ dbt_platform_helper/domain/terraform_environment.py
@@ -1,3 +1,4 @@
+from dbt_platform_helper.domain.versioning import PlatformHelperVersioning
 from dbt_platform_helper.platform_exception import PlatformException
 from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.providers.terraform_manifest import TerraformManifestProvider
@@ -19,11 +20,13 @@ class TerraformEnvironment:
         manifest_provider: TerraformManifestProvider = None,
         io: ClickIOProvider = ClickIOProvider(),
         platform_helper_version_status: PlatformHelperVersionStatus = PlatformHelperVersionStatus(),
+        platform_helper_versioning: PlatformHelperVersioning = PlatformHelperVersioning(),
     ):
         self.io = io
         self.config_provider = config_provider
         self.manifest_provider = manifest_provider or TerraformManifestProvider()
         self.platform_helper_version_status = platform_helper_version_status
+        self.platform_helper_versioning = platform_helper_versioning
 
     def generate(
         self,
@@ -37,17 +40,23 @@ class TerraformEnvironment:
                 f"cannot generate terraform for environment {environment_name}.  It does not exist in your configuration"
             )
 
-        platform_config_platform_helper_default_version = config.get("default_versions", {}).get(
-            "platform-helper"
-        )
+        # platform_config_platform_helper_default_version = config.get("default_versions", {}).get(
+        #     "platform-helper"
+        # )
 
-        self.platform_helper_version_status.cli_override = cli_platform_helper_version
-        self.platform_helper_version_status.platform_config_default = (
-            platform_config_platform_helper_default_version
-        )
+        # self.platform_helper_version_status.cli_override = cli_platform_helper_version
+        # self.platform_helper_version_status.platform_config_default = (
+        #     platform_config_platform_helper_default_version
+        # )
+
+        # platform_helper_version = (
+        #     self.platform_helper_version_status.get_required_platform_helper_version(self.io)
+        # )
+
+        self.platform_helper_versioning.cli_override = cli_platform_helper_version
 
         platform_helper_version = (
-            self.platform_helper_version_status.get_required_platform_helper_version(self.io)
+            self.platform_helper_versioning.get_required_platform_helper_version(self.io)
         )
 
         self.manifest_provider.generate_environment_config(
