+++ dbt_platform_helper/domain/versioning.py
@@ -1,5 +1,7 @@
 import os
+from typing import Optional
 
+from dbt_platform_helper.constants import MERGED_TPM_PLATFORM_HELPER_VERSION
 from dbt_platform_helper.platform_exception import PlatformException
 from dbt_platform_helper.providers.config import ConfigProvider
 from dbt_platform_helper.providers.io import ClickIOProvider
@@ -18,6 +20,7 @@ from dbt_platform_helper.providers.version import InstalledVersionProvider
 from dbt_platform_helper.providers.version import PyPiLatestVersionProvider
 from dbt_platform_helper.providers.version import VersionProvider
 from dbt_platform_helper.providers.version_status import PlatformHelperVersionStatus
+from dbt_platform_helper.providers.version_status import UnsupportedVersionException
 from dbt_platform_helper.providers.version_status import VersionStatus
 from dbt_platform_helper.providers.yaml_file import YamlFileProvider
 
@@ -46,6 +49,7 @@ class PlatformHelperVersioning:
         latest_version_provider: VersionProvider = PyPiLatestVersionProvider,
         installed_version_provider: InstalledVersionProvider = InstalledVersionProvider(),
         skip_versioning_checks: bool = None,
+        cli_override: Optional[SemanticVersion] = None,
     ):
         self.io = io
         self.version_file_version_provider = version_file_version_provider
@@ -55,6 +59,7 @@ class PlatformHelperVersioning:
         self.skip_versioning_checks = (
             skip_versioning_checks if skip_versioning_checks is not None else skip_version_checks()
         )
+        self.cli_override = cli_override
 
     def get_required_version(self, pipeline=None):
         version_status = self._get_version_status()
@@ -160,6 +165,36 @@ class PlatformHelperVersioning:
             raise PlatformHelperVersionNotFoundException
 
         return out
+
+    def get_required_platform_helper_version(self, io: ClickIOProvider):
+        version_status = self._get_version_status(include_project_versions=True)
+
+        version_preference_order = [
+            self.cli_override,
+            version_status.platform_config_default,
+            version_status.deprecated_version_file,
+        ]
+
+        # version_preference_order = [
+        #     self.cli_override,
+        #     self.platform_config_default,
+        #     self.deprecated_version_file,
+        # ]
+
+        valid_versions = [version for version in version_preference_order if version]
+
+        if valid_versions:
+            if SemanticVersion.is_semantic_version(valid_versions[0]):
+                semantic_version = SemanticVersion.from_string(valid_versions[0])
+                if semantic_version and (
+                    semantic_version.major < MERGED_TPM_PLATFORM_HELPER_VERSION
+                ):
+                    raise UnsupportedVersionException(valid_versions[0])
+            return valid_versions[0]
+        else:
+            io.warn(
+                "No platform-helper version specified. No value was provided via CLI, nor was one found in platform-config.yml under `default_versions`."
+            )
 
 
 class AWSVersioning:
