+++ dbt_platform_helper/providers/version_status.py
@@ -3,11 +3,9 @@ from dataclasses import field
 from typing import Dict
 from typing import Optional
 
-from dbt_platform_helper.constants import MERGED_TPM_PLATFORM_HELPER_VERSION
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
 from dbt_platform_helper.platform_exception import PlatformException
-from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.providers.semantic_version import SemanticVersion
 
 
@@ -90,24 +88,24 @@ class PlatformHelperVersionStatus(VersionStatus):
             "errors": errors,
         }
 
-    def get_required_platform_helper_version(self, io: ClickIOProvider):
-        version_preference_order = [
-            self.cli_override,
-            self.platform_config_default,
-            self.deprecated_version_file,
-        ]
+    # def get_required_platform_helper_version(self, io: ClickIOProvider):
+    #     version_preference_order = [
+    #         self.cli_override,
+    #         self.platform_config_default,
+    #         self.deprecated_version_file,
+    #     ]
 
-        valid_versions = [version for version in version_preference_order if version]
+    #     valid_versions = [version for version in version_preference_order if version]
 
-        if valid_versions:
-            if SemanticVersion.is_semantic_version(valid_versions[0]):
-                semantic_version = SemanticVersion.from_string(valid_versions[0])
-                if semantic_version and (
-                    semantic_version.major < MERGED_TPM_PLATFORM_HELPER_VERSION
-                ):
-                    raise UnsupportedVersionException(valid_versions[0])
-            return valid_versions[0]
-        else:
-            io.warn(
-                "No platform-helper version specified. No value was provided via CLI, nor was one found in platform-config.yml under `default_versions`."
-            )
+    #     if valid_versions:
+    #         if SemanticVersion.is_semantic_version(valid_versions[0]):
+    #             semantic_version = SemanticVersion.from_string(valid_versions[0])
+    #             if semantic_version and (
+    #                 semantic_version.major < MERGED_TPM_PLATFORM_HELPER_VERSION
+    #             ):
+    #                 raise UnsupportedVersionException(valid_versions[0])
+    #         return valid_versions[0]
+    #     else:
+    #         io.warn(
+    #             "No platform-helper version specified. No value was provided via CLI, nor was one found in platform-config.yml under `default_versions`."
+    #         )
